refactor(css_parser): Renamed 'CssSimpleFunction' to 'CssFunction'
All instances of 'CssSimpleFunction' have been replaced with 'CssFunction' across multiple files. This includes updates to function names, variable names, and type annotations.
