+++ crates/biome_css_formatter/src/css/auxiliary/mod.rs
@@ -22,6 +22,7 @@ pub(crate) mod declaration_list_block;
 pub(crate) mod declaration_or_at_rule_block;
 pub(crate) mod declaration_with_semicolon;
 pub(crate) mod document_custom_matcher;
+pub(crate) mod function;
 pub(crate) mod generic_delimiter;
 pub(crate) mod import_anonymous_layer;
 pub(crate) mod import_named_layer;
@@ -61,7 +62,6 @@ pub(crate) mod scope_edge;
 pub(crate) mod scope_range_end;
 pub(crate) mod scope_range_interval;
 pub(crate) mod scope_range_start;
-pub(crate) mod simple_function;
 pub(crate) mod supports_and_condition;
 pub(crate) mod supports_condition_in_parens;
 pub(crate) mod supports_feature_declaration;
