+++ crates/biome_css_parser/src/syntax/parse_error.rs
@@ -43,19 +43,6 @@ pub(crate) fn expected_non_css_wide_keyword_identifier(
     }
 }
 
-pub(crate) fn expected_expression(p: &CssParser, range: TextRange) -> ParseDiagnostic {
-    expected_any(
-        &[
-            "component value",
-            "binary expression",
-            "parenthesized expression",
-            "any function expression",
-        ],
-        range,
-        p,
-    )
-}
-
 pub(crate) fn expected_number(p: &CssParser, range: TextRange) -> ParseDiagnostic {
     expected_node("number", range, p)
 }
