+++ crates/biome_css_parser/src/syntax/value/mod.rs
@@ -1,2 +1,4 @@
+pub(crate) mod dimension;
+pub(crate) mod function;
 mod parse_error;
 pub(crate) mod url;
